Add win/loss screen freeze
Instead of immediately initializing the new game screen, the last
frame will be shown to help the user understand what was going on.
Also, instead of continuing with any key press, return is required
to avoid accidental skipping of this screen.
Along with this change, escape can now be used while in the
between-rounds-screen.