Add a game snapshot function
Add a function that delivers a snapshot of a the game to be able
to display it to the user in preparation. It is called with an
initial ships array and returns a world independent of the
actual (initialized) game world.
This is rather complicated to make it work for all parameter
combinations.