Improve status reports
Completely change the way the timeout, next action, ship life and
ship ammunition are displayed to make it easier to  understand
and faster to gather while gaming.
Also add an indication when on low life.
It should be easier to understand now why a ship gets wracked if
you move it with one life.